minor fixes for test page
