+++ src/pages/MonthlyGood/EmailPreviewContentful.vue
@@ -1,17 +1,13 @@
 <template>
 	<div class="email-preview-section-wrapper row">
 		<div class="small-12 large-6 columns">
-			<!-- <picture>
-				<source srcset="bootstrap-extra.jpg" media="(min-width: 1200px)" />
-				<source srcset="bootstrap-large.jpg" media="(min-width: 992px)" />
-				<source srcset="bootstrap-medium.jpg " media="(min-width: 768px)" />
-				<source srcset="bootstrap-small.jpg" media="(min-width: 576px)" />
-				<img src="bootstrap-extra-small.jpg" class="img-fluid" />
-			</picture> -->
+			<!-- responsive image based on BS breakpoints -->
+			<!-- Values are set based on 'small-12 large-6' values above -->
+			<!-- TODO use KvResponsiveImage2 or create new component -->
 			<picture v-for="(media, index) in contentGroup.media" :key="index">
-				<source media="(min-width: 800px)" :srcset="media.fields.file.url +'?w=800&h=600&fm=webp&q=80'">
-				<source media="(min-width: 400px)" :srcset="media.fields.file.url + '?w=400&h=300&fm=webp&q=80'">
-				<img :src="media.fields.file.url" :alt="media.fields.description" style="width: auto;">
+				<source media="(min-width: 681px)" :srcset="media.fields.file.url + '?w=495&fm=webp&q=80'">
+				<source media="(min-width: 400px)" :srcset="media.fields.file.url + '?w=400&fm=webp&q=80'">
+				<img :src="media.fields.file.url" :alt="media.fields.description">
 			</picture>
 		</div>
 		<div class="small-12 large-6 columns">
@@ -19,18 +15,16 @@
 				<h2 class="impact-text">
 					{{ contentGroup.content.fields.headline }}
 				</h2>
-				<div v-html="bodyCopy">
-				</div>
+				<div v-html="bodyCopy"></div>
 			</div>
 		</div>
 	</div>
 </template>
 
 <script>
-// import KvResponsiveImage from '@/components/Kv/KvResponsiveImage';
 import { documentToHtmlString } from '~/@contentful/rich-text-html-renderer';
 
-const emailPreviewImageRequire = require.context('@/assets/images/mg-email-preview', true);
+// const emailPreviewImageRequire = require.context('@/assets/images/mg-email-preview', true);
 
 export default {
 	props: {
@@ -41,25 +35,21 @@ export default {
 			}
 		},
 	},
-	components: {
-		// KvResponsiveImage,
-	},
 	computed: {
 		bodyCopy() {
 			return documentToHtmlString(this.contentGroup.content.fields.bodyCopy);
 		}
 	},
-	data() {
-		return {
-			images: [
-				['small', emailPreviewImageRequire('./email-preview-mobile-std.jpg')],
-				['small retina', emailPreviewImageRequire('./email-preview-mobile-retina.jpg')],
-				['xga', emailPreviewImageRequire('./email-preview-desktop-std.jpg')],
-				['xga retina', emailPreviewImageRequire('./email-preview-desktop-retina.jpg')],
-			]
-		};
-	},
-
+	// data() {
+	// 	return {
+	// 		images: [
+	// 			['small', emailPreviewImageRequire('./email-preview-mobile-std.jpg')],
+	// 			['small retina', emailPreviewImageRequire('./email-preview-mobile-retina.jpg')],
+	// 			['xga', emailPreviewImageRequire('./email-preview-desktop-std.jpg')],
+	// 			['xga retina', emailPreviewImageRequire('./email-preview-desktop-retina.jpg')],
+	// 		]
+	// 	};
+	// },
 };
 </script>
 
@@ -76,7 +66,8 @@ h2 {
 }
 
 img {
-	max-width: 21.88rem;
+	max-width: 100%;
+	height: auto;
 	margin: 0 auto;
 	display: block;
 }
