+++ src/pages/MonthlyGood/KivaAsExpertContentful.vue
@@ -7,29 +7,28 @@
 				</h2>
 			</div>
 			<div class="small-12 large-7 columns">
-				<div v-html="bodyCopy">
-				</div>
+				<div v-html="bodyCopy"></div>
 				<slot name="form"></slot>
 			</div>
 
 			<div class="large-5 columns">
-				<!-- {{ contentGroup.media.fields.file.url }} -->
-				<img v-for="(media, index) in contentGroup.media" :src="media.fields.file.url" :key="index">
-				<!-- <kv-responsive-image
-					:images="billionImpactImages"
-					class="show-for-large community-image"
-					alt="10 years and $1 billion in loans funded"
-				/> -->
+				<!-- responsive image based on BS breakpoints -->
+				<!-- Values are set based on 'small-12 large-6' values above -->
+				<!-- TODO use KvResponsiveImage2 or create new component -->
+				<picture v-for="(media, index) in contentGroup.media" :key="index">
+					<source media="(min-width: 681px)" :srcset="media.fields.file.url + '?w=495&fm=webp&q=80'">
+					<source media="(min-width: 400px)" :srcset="media.fields.file.url + '?w=400&fm=webp&q=80'">
+					<img :src="media.fields.file.url" :alt="media.fields.description">
+				</picture>
 			</div>
 		</div>
 	</div>
 </template>
 
 <script>
-// import KvResponsiveImage from '@/components/Kv/KvResponsiveImage';
 import { documentToHtmlString } from '~/@contentful/rich-text-html-renderer';
 
-const billionImpactImagesRequire = require.context('@/assets/images/10-years-billion-impact', true);
+// const billionImpactImagesRequire = require.context('@/assets/images/10-years-billion-impact', true);
 
 export default {
 	props: {
@@ -40,17 +39,14 @@ export default {
 			}
 		},
 	},
-	data() {
-		return {
-			billionImpactImages: [
-				['small', billionImpactImagesRequire('./10-years-billion-impact_ghost.jpg')],
-				['small retina', billionImpactImagesRequire('./10-years-billion-impact_2x_ghost.jpg')],
-			]
-		};
-	},
-	components: {
-		// KvResponsiveImage
-	},
+	// data() {
+	// 	return {
+	// 		billionImpactImages: [
+	// 			['small', billionImpactImagesRequire('./10-years-billion-impact_ghost.jpg')],
+	// 			['small retina', billionImpactImagesRequire('./10-years-billion-impact_2x_ghost.jpg')],
+	// 		]
+	// 	};
+	// },
 	computed: {
 		bodyCopy() {
 			return documentToHtmlString(this.contentGroup.content.fields.bodyCopy);
